+++ commons/src/lib.rs
@@ -2,10 +2,6 @@ pub mod cache;
 pub mod display;
 pub mod gem_version;
 pub mod gemfile_lock;
-#[deprecated(
-    since = "0.0.0",
-    note = "Use the struct layer API in the latest libcnb.rs instead"
-)]
 pub mod layer;
 pub mod metadata_digest;
 pub mod output;